requirements.txt
langchain==0.3.14
langchain-core==0.3.30
langchain-cohere==0.3.4
python-dotenv==1.0.1
